pokemon_summary_screen.c, part 2.
Also fix prototype of MenuHelpers_CallLinkSomething.

+++ src/item_pc.c
@@ -379,7 +379,7 @@ static bool8 ItemPc_DoGfxSetup(void)
         gMain.state++;
         break;
     case 20:
-        if ((u8)MenuHelpers_CallLinkSomething() != TRUE)
+        if (MenuHelpers_CallLinkSomething() != TRUE)
             gMain.state++;
         break;
     default:
